- Changed Opentracker build system to the GNU configure machinery. - renamed config.h to config_win.h and made the appropriate changes in the src files   in regard to windows. - cleanups for gcc3 to compile without warnings.   (gcc3 is agreed to be the default compiler for unix - so no more #ifdef's needed to    handle special #include cases) - configure options:   checks for curses by default - can let it work for any unix curses implementation   with some #ifdef's if ever needed (ncurses is our friend!)   --enable-speech         enable the Speech recognition (default=no)   --enable-joystick       enable the use of joysticks (default=no)   --with-ace=DIR          give prefix location of ACE   --with-xerces=DIR       give prefix location of Xerces   --with-artoolkit=DIR    give prefix location of ARToolkit (<- this is optional but                                                              issues a warning if missing) - added opentracker_env.sh to allow easy setting of the environment variables   call it like . ./opentracker_env.sh from your shell - configure style is like the studierstube intree. no install targets have been   specified. (i.e. no need to call make install afterwards, you work in the   source tree and the lib gets built in ./lib) - intersense driver for unix is built in by default as part of libOpenTracker - after a succesful build you will find 2 test applications in ./bin (main && test) - have fun! :)

+++ src/core/OSUtils.cxx
@@ -26,7 +26,7 @@
   *
   * @author Gerhard Reitmayr
   *
-  * $Header: /scratch/subversion/cvs2svn-0.1236/../cvs/opentracker/src/core/OSUtils.cxx,v 1.4 2002/11/08 21:03:13 reitmayr Exp $
+  * $Header: /scratch/subversion/cvs2svn-0.1236/../cvs/opentracker/src/core/OSUtils.cxx,v 1.5 2003/01/09 04:14:12 tamer Exp $
   * @file                                                                   */
  /* ======================================================================= */
 
@@ -42,7 +42,7 @@ double OSUtils::currentTime()
 //sleeps the specified amount of time ...
 void OSUtils::sleep( double time )
 {
-    ACE_Time_Value timeVal(0, 1000*time );
+    ACE_Time_Value timeVal(0, (int)(1000*time) );
 	ACE_OS::sleep( timeVal );
 }
